disabled plugin: render-markdown due to alignment problems
Need to figure out how to add more forced whitespace to the left of the
buffer to account for the varying number of columns that are displayed
when going between Normal and Insert mode.

Normal mode inserts certain special characters to indicate a markdown
tag, then Insert takes that character away. This makes the width of the
buffer/editor vary between nodes and it's jarring to switch between.